Add files via upload
This is the first project I completed in my data science journey. It was an interesting project which dives into how  several factors  affected the destruction of Titanic's maiden voyage and resulted in one of the biggest tragedies human kind ever witnessed. Through this EDA we delve deeply into how different factors played a crucial role in the survival of its passengers.